Add icons instead of buttons
So this commit is basically just me testing the usage of actual images or icons if you will, for the different choices of weapons since its more engaging and looks better overall. I got the whole functionality working for now so it should be easy to add the other 2 weapons.