Remove game() function and commented out function
So did quite some stuff here! First, I removed the game() function because it was just there sitting and doing nothing since we are not using the console anymore!

Next, I removed the commented out code for the function which worked with the buttons!

Finally, hit a bit of a block road but it was all easy to fix. INitially the challenge was adding a way to identify each image with a value, let this value be either 'Rock', 'Paper' or 'scissors'. The solution was simply to add an ID to each image and then target it using event.target.id.